backend/requirements.txt
fastapi
uvicorn[standard]
onnxruntime
Pillow
numpy




